Create ft1
In the above example, the EtherAcceptingContract contract includes a receive() function with the external and payable modifiers. This function acts as a fallback function for receiving Ether. When someone sends Ether to this contract, the receive() function will be triggered automatically, and the contract's balance will be increased accordingly.

The contract also includes a fallback() function with the external modifier. This function serves as the fallback function for any unrecognized calls or transactions. If someone calls a function that doesn't exist in the contract or if the function signature doesn't match any existing functions, the fallback() function will be executed.